Adding a "paddingTop" option
I made this update in the code because I needed to add some paddingTop for the plugin. I have a fixed menu and I needed to have this option to compensate the menu height.
So, I just added 2 lines of code to make it works.
It's really simple, but I think it would be usefull for the plugin and everyone who use it.